requirements.txt
altair==5.5.0
altgraph @ file:///AppleInternal/Library/BuildRoots/4~CKgHugAcJujzKlRoChleiLcli8Pvfh2Zw9NQTKg/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
attrs==26.1.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.1.8
coloredlogs==15.0.1
flatbuffers==25.12.19
future @ file:///AppleInternal/Library/BuildRoots/4~CKgHugAcJujzKlRoChleiLcli8Pvfh2Zw9NQTKg/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
gitdb==4.0.12
GitPython==3.1.46
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy-loader==0.5
llvmlite==0.43.0
macholib @ file:///AppleInternal/Library/BuildRoots/4~CKgHugAcJujzKlRoChleiLcli8Pvfh2Zw9NQTKg/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==3.0.3
mpmath==1.3.0
narwhals==2.19.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
onnxruntime==1.19.2
opencv-python-headless==4.13.0.92
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.4.0
pooch==1.9.0
protobuf==6.33.6
pyarrow==21.0.0
pydeck==0.9.1
PyMatting==1.1.15
python-dateutil==2.9.0.post0
pytz==2026.1.post1
referencing==0.36.2
rembg==2.0.61
requests==2.32.5
rpds-py==0.27.1
scikit-image==0.24.0
scipy==1.13.1
six @ file:///AppleInternal/Library/BuildRoots/4~CKgHugAcJujzKlRoChleiLcli8Pvfh2Zw9NQTKg/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
smmap==5.0.3
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
tifffile==2024.8.30
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
